Move process_pid functionality into shared
